:recycle: Refactor: MapProivderItem 컴포저블 분리
