+++ composeApp/src/commonMain/kotlin/com/dh/ondot/presentation/home/HomeScreen.kt
@@ -2,8 +2,6 @@ package com.dh.ondot.presentation.home
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -15,7 +13,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -24,7 +21,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -36,20 +32,16 @@ import com.dh.ondot.presentation.home.components.EmptyScheduleContent
 import com.dh.ondot.presentation.home.components.RemainingTimeText
 import com.dh.ondot.presentation.home.components.ScheduleList
 import com.dh.ondot.presentation.home.components.UserBadgeBanner
+import com.dh.ondot.presentation.ui.components.MapProviderItem
 import com.dh.ondot.presentation.ui.components.OnDotText
 import com.dh.ondot.presentation.ui.theme.CREATE_SCHEDULE_GUIDE
 import com.dh.ondot.presentation.ui.theme.MAP_PROVIDER_CONTENT
 import com.dh.ondot.presentation.ui.theme.MAP_PROVIDER_TITLE
 import com.dh.ondot.presentation.ui.theme.OnDotColor.Gray0
-import com.dh.ondot.presentation.ui.theme.OnDotColor.Gray400
 import com.dh.ondot.presentation.ui.theme.OnDotColor.Gray600
 import com.dh.ondot.presentation.ui.theme.OnDotColor.Gray900
-import com.dh.ondot.presentation.ui.theme.OnDotColor.Green600
 import ondot.composeapp.generated.resources.Res
-import ondot.composeapp.generated.resources.ic_apple_map
 import ondot.composeapp.generated.resources.ic_banner
-import ondot.composeapp.generated.resources.ic_kakao_map
-import ondot.composeapp.generated.resources.ic_naver_map
 import org.jetbrains.compose.resources.painterResource
 
 @Composable
@@ -219,49 +211,3 @@ private fun MapProviderDialog(
     }
 }
 
-@Composable
-private fun MapProviderItem(
-    provider: MapProvider,
-    isSelected: Boolean,
-    modifier: Modifier = Modifier,
-    interactionSource: MutableInteractionSource,
-    onClick: () -> Unit
-) {
-    val imageResource = when(provider) {
-        MapProvider.KAKAO -> painterResource(Res.drawable.ic_kakao_map)
-        MapProvider.NAVER -> painterResource(Res.drawable.ic_naver_map)
-        MapProvider.APPLE -> painterResource(Res.drawable.ic_apple_map)
-    }
-
-    Box(
-        modifier = modifier
-            .background(Gray400, RoundedCornerShape(12.dp))
-            .clip(RoundedCornerShape(12.dp))
-            .border(1.dp, if (isSelected) Green600 else Gray400, RoundedCornerShape(12.dp))
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                onClick = onClick
-            )
-            .padding(12.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Image(
-                painter = imageResource,
-                contentDescription = null,
-                modifier = Modifier.size(46.dp)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            OnDotText(
-                text = provider.providerName,
-                style = OnDotTextStyle.BodyMediumR,
-                color = Gray0
-            )
-        }
-    }
-}
